Made a rudimentary enemy AI
Fixed some minor bugs
Added in an enemy AI: for each move it can do it will evaluate how good that move is using movePriority, and then do a random move of the highest priority.
movePriority currently just takes in whether it puts the enemy king in check and what pieces it kills, and does not look ahead into the future.